+++ openrouteservice/src/main/java/org/heigit/ors/routing/RouteSearchParameters.java
@@ -81,6 +81,9 @@ public class RouteSearchParameters {
     private int roundTripPoints = 2;
     private long roundTripSeed = -1;
 
+    private double maximumSpeed;
+    private boolean hasMaximumSpeed = false;
+
     private String options;
 
     public int getProfileType() {
@@ -531,6 +534,19 @@ public class RouteSearchParameters {
         return roundTripSeed;
     }
 
+    public double getMaximumSpeed() {
+        return maximumSpeed;
+    }
+
+    public void setMaximumSpeed(double maximumSpeed) {
+        this.maximumSpeed = maximumSpeed;
+        hasMaximumSpeed = true;
+    }
+
+    public boolean hasMaximumSpeed() {
+        return hasMaximumSpeed;
+    }
+
     public boolean isProfileTypeDriving() {
         return RoutingProfileType.isDriving(this.getProfileType());
     }
@@ -547,7 +563,7 @@ public class RouteSearchParameters {
             || getConsiderTurnRestrictions()
             || isProfileTypeHeavyVehicle() && getVehicleType() > 0
             || isProfileTypeDriving() && hasParameters(VehicleParameters.class)
-        ;
+            || hasMaximumSpeed();
     }
 
     /**