+++ openrouteservice/src/main/java/org/heigit/ors/routing/RoutingProfile.java
@@ -890,6 +890,10 @@ public class RoutingProfile {
                 req.getHints().put("alternative_route.max_share_factor", searchParams.getAlternativeRoutesShareFactor());
             }
 
+            if(searchParams.hasMaximumSpeed()){
+                req.getHints().put("maximum_speed", searchParams.getMaximumSpeed());
+            }
+
             if (directedSegment) {
                 resp = mGraphHopper.constructFreeHandRoute(req);
             } else {