+++ openrouteservice/src/main/java/org/heigit/ors/routing/configuration/RouteProfileConfiguration.java
@@ -57,6 +57,8 @@ public class RouteProfileConfiguration {
 	private int locationIndexResolution = 500;
 	private int locationIndexSearchIterations = 4;
 
+	private double maximumSpeedLowerBound = 80;
+
 	public RouteProfileConfiguration() {
 		extStorages = new HashMap<>();
 		graphBuilders = new HashMap<>();
@@ -93,6 +95,8 @@ public class RouteProfileConfiguration {
 		hasMaximumSnappingRadius = rpc.hasMaximumSnappingRadius;
 
 		extent = rpc.extent;
+
+		maximumSpeedLowerBound = rpc.maximumSpeedLowerBound;
 	}
 
 	public Integer[] getProfilesTypes() {
@@ -365,5 +369,13 @@ public class RouteProfileConfiguration {
 
 	public void setLocationIndexSearchIterations(int locationIndexSearchIterations) {
 		this.locationIndexSearchIterations = locationIndexSearchIterations;
+	}
+
+	public void setMaximumSpeedLowerBound(double maximumSpeedLowerBound){
+		this.maximumSpeedLowerBound = maximumSpeedLowerBound;
+	}
+
+	public double getMaximumSpeedLowerBound(){
+		return maximumSpeedLowerBound;
 	}
 }